HLS intersect: Split header files to sw-only and common (for both)
