+++ actions/hls_intersect/README.md
@@ -4,7 +4,8 @@ Intersection has two methods and are implemented in two Actions.
 One is in hw_h (hash, -m1) and one is in hw_s (sort, -m2).
 You must configure it first (otherwise the default one is 'hash').
 
-Under $SNAP_ROOT/hardware directory
+After `make snap_config`, do this perticular step. (In case $ACTION_ROOT is not set, try to source snap_path.sh)
+
 ```
 make clean
 
@@ -16,9 +17,9 @@ make -C $ACTION_ROOT config_s
 
 Then follow the normal flow
 ``` 
-make config model
+make model
  .or. 
-make config image
+make image
 ```
